requirements.txt
numpy~=1.24.3
PyYAML~=5.3.1

Pillow~=7.0.0